Fixed the issue related to floating point numbers

+++ demo/src/main/java/com/epam/demo/Utils.java
@@ -5,6 +5,6 @@ import java.util.List;
 public class Utils {
     public static boolean isAllPositiveNumbers(List<String> args) {
         return args.stream()
-                .allMatch(str -> Integer.parseInt(str) > 0);
+                .allMatch(str -> Double.parseDouble(str) > 0);
     }
 }
